Split widget data store tests by concern
